% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_vert_traitDB.R
\name{combine_vert_traitDB}
\alias{combine_vert_traitDB}
\title{Function to combine trait databases on mammals, birds, reptiles and amphibians of Europe.}
\usage{
combine_vert_traitDB()
}
\value{
A database with common traits from the four groups. See W. Thuiller et al., 2015, Philosophical Transactions of the Royal Society B: Biological Sciences
for details on trait identification
}
\description{
Function to combine trait databases on mammals, birds, reptiles and amphibians of Europe.
}
