% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genus.R
\name{get_genus}
\alias{get_genus}
\title{Get genus level from a list of species name}
\usage{
get_genus(species.vec)
}
\arguments{
\item{species.vec}{vector of species name}
}
\value{
Vector of genius associated to species name
}
\description{
Get genus level from a list of species name
}
