% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_disperse.R
\name{prep_disperse}
\alias{prep_disperse}
\title{Transformation of DISPERSE general database into a unique trait-species relationship}
\usage{
prep_disperse()
}
\description{
Transformation of DISPERSE general database into a unique trait-species relationship
}
