% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_movmode.R
\name{prep_movmode}
\alias{prep_movmode}
\title{Transformation of TAXREF mobility file into a binary database}
\usage{
prep_movmode()
}
\description{
Transformation of TAXREF mobility file into a binary database
}
