% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_tetra.R
\name{prep_tetra}
\alias{prep_tetra}
\title{Transformation of TETRA_EU general diet file into a binary database}
\usage{
prep_tetra()
}
\value{
A dataframe with each line being one species and columns indicating whether the species eats what is specified in column name
}
\description{
Transformation of TETRA_EU general diet file into a binary database
}
