% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileTraits-package.R
\docType{package}
\name{CompileTraits-package}
\alias{CompileTraits}
\alias{CompileTraits-package}
\title{CompileTraits: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mcpri3/CompileTraits}
  \item Report bugs at \url{https://github.com/mcpri3/CompileTraits/issues}
}

}
\author{
\strong{Maintainer}: Marie-Caroline Prima \email{marie-caroline.prima@univ-grenoble-alpes.fr} [copyright holder]

}
\keyword{internal}
